async_hooks: deprecate undocumented API
