+++ lib/internal/bootstrap_node.js
@@ -401,7 +401,7 @@
         // Emit the after() hooks now that the exception has been handled.
         if (async_hook_fields[kAfter] > 0) {
           do {
-            NativeModule.require('async_hooks').emitAfter(
+            NativeModule.require('internal/async_hooks').emitAfter(
               async_id_fields[kExecutionAsyncId]);
           } while (asyncIdStackSize() > 0);
         // Or completely empty the id stack.
