+++ lib/internal/process/next_tick.js
@@ -48,7 +48,7 @@ class NextTickQueue {
 
 function setupNextTick() {
   const async_wrap = process.binding('async_wrap');
-  const async_hooks = require('async_hooks');
+  const async_hooks = require('internal/async_hooks');
   const promises = require('internal/process/promises');
   const errors = require('internal/errors');
   const emitPendingUnhandledRejections = promises.setup(scheduleMicrotasks);
