+++ lib/timers.js
@@ -40,7 +40,7 @@ const {
   emitBefore,
   emitAfter,
   emitDestroy
-} = require('async_hooks');
+} = require('internal/async_hooks');
 // Grab the constants necessary for working with internal arrays.
 const { kInit, kDestroy, kAsyncIdCounter } = async_wrap.constants;
 // Symbols for storing async id state.
