+++ test/parallel/test-async-hooks-run-in-async-id-scope.js
@@ -4,7 +4,7 @@ const common = require('../common');
 const assert = require('assert');
 const async_hooks = require('async_hooks');
 
-const asyncId = async_hooks.newUid();
+const asyncId = new async_hooks.AsyncResource('test').asyncId();
 
 assert.notStrictEqual(async_hooks.executionAsyncId(), asyncId);
